backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic==2.9.2
pydantic-settings==2.5.2
pymongo[srv]==4.7.3
python-dotenv==1.0.1
httpx==0.27.0
requests>=2.31.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-mongodb>=0.1.0
langgraph>=0.2.40
langsmith>=0.1.17
tiktoken>=0.7.0
elevenlabs>=1.0.0
twilio>=9.0.0

